Add OS X support to helper package.
Simply reuse the linux version for both linux and darwin via build
constraints.
