+++ examples/opengl-tutorial/helper/helper.go
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// +build linux darwin
+
 package helper
 
 import (
